Addressed review points.
